# docs/04_API/OpenAPI_V1.yaml
openapi: 3.0.3
info:
  title: ZEROON API
  version: 1.0.0
  description: Contract for the ZEROON MVP mobile app and admin system.
  license:
    name: Proprietary
    url: https://zeroon.ai/legal/license
servers:
  - url: https://api.zeroon.ai/api/v1
    description: Production
...
paths:
...
  /support/requests:
    post:
      tags: [Support]
      summary: Submit a private support request
      description: >-
        Creates one owner-scoped request with an opaque public reference.
        Retrying the same clientSubmissionId and identical body returns the
        existing request; reusing the id with different content returns 409.
        Diagnostics are optional, explicitly consented, and allowlisted. No
        Record, Memory, Profile, conversation, prompt, reply, token, secret,
        screenshot, attachment, or full log is accepted.
      operationId: createSupportRequest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportRequest"
      responses:
        "201":
          description: Support request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportRequestDetail"
        "200":
          description: Existing support request returned for an identical idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportRequestDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/SupportRateLimited"
  /me/support-requests:
    get:
      tags: [Support]
      summary: List support requests owned by the current user
      operationId: listCurrentUserSupportRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/SupportSize"
      responses:
        "200":
          description: Owned support request page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportRequestPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/support-requests/{reference}:
    get:
      tags: [Support]
      summary: Get one support request owned by the current user
      operationId: getCurrentUserSupportRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SupportReference"
      responses:
        "200":
          description: Owned support request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportRequestDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/support-requests/{reference}/messages:
    post:
      tags: [Support]
      summary: Add a follow-up to one owned support request
      description: >-
        Adds a bounded user-visible message. CLOSED requests reject follow-up.
        A follow-up to WAITING_FOR_USER or REPLIED returns the request to
        IN_REVIEW and records the transition.
      operationId: addCurrentUserSupportMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SupportReference"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSupportMessageRequest"
      responses:
        "201":
          description: Follow-up stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/SupportRateLimited"
  /admin/support-requests:
    get:
      tags: [Support]
      summary: List support requests for human operator handling
      description: >-
        ADMIN-only metadata queue. It never includes unrelated ZEROON private
        content, internal database ids, or message bodies.
      operationId: listAdminSupportRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/SupportSize"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/SupportRequestStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/SupportCategory"
        - name: escalated
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Filtered operator queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSupportRequestPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/support-requests/{reference}:
    get:
      tags: [Support]
      summary: Read one support request for human handling
      operationId: getAdminSupportRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SupportReference"
      responses:
        "200":
          description: Operator detail including explicitly separated internal notes and audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSupportRequestDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Support]
      summary: Apply bounded triage changes to one support request
      description: >-
        ADMIN may change category, assign or unassign themself, set or clear a
        reviewed escalation, and take a RECEIVED request into IN_REVIEW. Every
        effective mutation creates a content-free audit row. Closing and other
        user-visible status changes require the reply operation.
      operationId: updateAdminSupportRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SupportReference"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSupportUpdateRequest"
      responses:
        "200":
          description: Updated operator detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSupportRequestDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/support-requests/{reference}/messages:
    post:
      tags: [Support]
      summary: Add an internal note or audited user-visible human reply
      description: >-
        INTERNAL is visible only to ADMIN and cannot change status.
        USER_VISIBLE requires one allowed next status and is stored atomically
        with its transition. Closing therefore always includes a visible
        explanation. Closed requests cannot be mutated or reopened.
      operationId: addAdminSupportMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SupportReference"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSupportMessageRequest"
      responses:
        "201":
          description: Message, transition, and audit committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSupportRequestDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
    SupportSize:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    SupportReference:
      name: reference
      in: path
      required: true
      schema:
        type: string
        pattern: "^ZS-[A-F0-9]{20}$"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Insufficient permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Idempotency conflict or request state does not allow the operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    SupportRateLimited:
      description: Support request or follow-up rate limit exceeded
      headers:
        Retry-After:
          description: Minimum seconds before retrying this operation
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: rate_limited
        message:
          type: string
...
    SupportCategory:
      type: string
      enum:
        - PRODUCT_PROBLEM
        - SUGGESTION
        - ACCOUNT_DATA_PRIVACY
        - AI_RESPONSE_SAFETY
        - COMPLAINT_RIGHTS
        - OTHER
    SupportRequestStatus:
      type: string
      enum: [RECEIVED, IN_REVIEW, WAITING_FOR_USER, REPLIED, CLOSED]
    SupportActorType:
      type: string
      enum: [SYSTEM, USER, ADMIN]
    SupportMessageVisibility:
      type: string
      enum: [USER_VISIBLE, INTERNAL]
    SupportEscalationCode:
      type: string
      enum:
        - ENGINEERING
        - PRODUCT
        - PRIVACY
        - SAFETY
        - COMPLAINT
        - PROFESSIONAL_REVIEW
    SupportAdminAuditAction:
      type: string
      enum:
        - CATEGORY_CHANGE
        - ASSIGNMENT_CHANGE
        - ESCALATION_CHANGE
        - STATUS_CHANGE
        - USER_VISIBLE_REPLY
        - INTERNAL_NOTE
    CreateSupportRequest:
      type: object
      additionalProperties: false
      required:
        [clientSubmissionId, category, subject, description, diagnosticConsent]
      properties:
        clientSubmissionId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/SupportCategory"
        subject:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 4000
        replyContact:
          type: string
          maxLength: 200
          nullable: true
        diagnosticConsent:
          type: boolean
          description: Must be true when diagnostics is present.
        diagnostics:
          $ref: "#/components/schemas/SupportDiagnosticEnvelope"
          nullable: true
    SupportDiagnosticEnvelope:
      type: object
      additionalProperties: false
      properties:
        appVersion:
          type: string
          maxLength: 40
        build:
          type: string
          maxLength: 40
        platform:
          type: string
          maxLength: 30
        osFamily:
          type: string
          maxLength: 40
        locale:
          type: string
          enum: [zh-CN, en]
        errorCode:
          type: string
          maxLength: 80
          pattern: "^[A-Za-z0-9._:-]*$"
        timestamp:
          type: string
          format: date-time
    AddSupportMessageRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
    SupportRequestSummary:
      type: object
      required: [reference, category, status, subject, createdAt, updatedAt]
      properties:
        reference:
          type: string
          pattern: "^ZS-[A-F0-9]{20}$"
        category:
          $ref: "#/components/schemas/SupportCategory"
        status:
          $ref: "#/components/schemas/SupportRequestStatus"
        subject:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SupportRequestPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SupportRequestSummary"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 50
        totalElements:
          type: integer
          format: int64
          minimum: 0
    SupportMessage:
      type: object
      required: [id, actorType, body, createdAt]
      properties:
        id:
          type: integer
          format: int64
        actorType:
          $ref: "#/components/schemas/SupportActorType"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    SupportStatusHistory:
      type: object
      required: [toStatus, actorType, createdAt]
      properties:
        fromStatus:
          $ref: "#/components/schemas/SupportRequestStatus"
          nullable: true
        toStatus:
          $ref: "#/components/schemas/SupportRequestStatus"
        actorType:
          $ref: "#/components/schemas/SupportActorType"
        createdAt:
          type: string
          format: date-time
    SupportRequestDetail:
      allOf:
        - $ref: "#/components/schemas/SupportRequestSummary"
        - type: object
          required: [description, messages, statusHistory]
          properties:
            description:
              type: string
            replyContact:
              type: string
              nullable: true
            diagnostics:
              $ref: "#/components/schemas/SupportDiagnosticEnvelope"
              nullable: true
            messages:
              type: array
              items:
                $ref: "#/components/schemas/SupportMessage"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/SupportStatusHistory"
            closedAt:
              type: string
              format: date-time
              nullable: true
    AdminSupportUpdateRequest:
      type: object
      additionalProperties: false
      required: [reasonCode]
      minProperties: 2
      properties:
        category:
          $ref: "#/components/schemas/SupportCategory"
        status:
          $ref: "#/components/schemas/SupportRequestStatus"
          description: Status-only mutation is limited to RECEIVED to IN_REVIEW.
        assignToMe:
          type: boolean
          description: True assigns the authenticated ADMIN; false unassigns.
        escalated:
          type: boolean
        escalationCode:
          $ref: "#/components/schemas/SupportEscalationCode"
          nullable: true
          description: Required when escalated is true and absent when false.
        reasonCode:
          type: string
          pattern: "^[A-Z0-9_]{2,40}$"
          maxLength: 40
    AdminSupportMessageRequest:
      type: object
      additionalProperties: false
      required: [body, visibility, reasonCode]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
        visibility:
          $ref: "#/components/schemas/SupportMessageVisibility"
        nextStatus:
          $ref: "#/components/schemas/SupportRequestStatus"
          nullable: true
          description: Required for USER_VISIBLE and forbidden for INTERNAL.
        reasonCode:
          type: string
          pattern: "^[A-Z0-9_]{2,40}$"
          maxLength: 40
    AdminSupportRequestSummary:
      type: object
      required:
        - reference
        - ownerUid
        - category
        - status
        - subject
        - descriptionPreview
        - escalated
        - createdAt
        - updatedAt
      properties:
        reference:
          type: string
          pattern: "^ZS-[A-F0-9]{20}$"
        ownerUid:
          type: string
        category:
          $ref: "#/components/schemas/SupportCategory"
        status:
          $ref: "#/components/schemas/SupportRequestStatus"
        subject:
          type: string
        descriptionPreview:
          type: string
          maxLength: 160
        assignedAdminUid:
          type: string
          nullable: true
        escalated:
          type: boolean
        escalationCode:
          $ref: "#/components/schemas/SupportEscalationCode"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminSupportRequestPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminSupportRequestSummary"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 50
        totalElements:
          type: integer
          format: int64
          minimum: 0
    AdminSupportMessage:
      type: object
      required: [id, actorType, visibility, body, createdAt]
      properties:
        id:
          type: integer
          format: int64
        actorType:
          $ref: "#/components/schemas/SupportActorType"
        visibility:
          $ref: "#/components/schemas/SupportMessageVisibility"
        actorUid:
          type: string
          nullable: true
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminSupportStatusHistory:
      type: object
      required: [toStatus, actorType, reasonCode, createdAt]
      properties:
        fromStatus:
          $ref: "#/components/schemas/SupportRequestStatus"
          nullable: true
        toStatus:
          $ref: "#/components/schemas/SupportRequestStatus"
        actorType:
          $ref: "#/components/schemas/SupportActorType"
        actorUid:
          type: string
          nullable: true
        reasonCode:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminSupportAudit:
      type: object
      required: [actionType, reasonCode, createdAt]
      description: Structured mutation evidence that never copies message or request body text.
      properties:
        actionType:
          $ref: "#/components/schemas/SupportAdminAuditAction"
        actorUid:
          type: string
          nullable: true
        fromValue:
          type: string
          nullable: true
        toValue:
          type: string
          nullable: true
        messageId:
          type: integer
          format: int64
          nullable: true
        reasonCode:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminSupportRequestDetail:
      allOf:
        - $ref: "#/components/schemas/AdminSupportRequestSummary"
        - type: object
          required: [description, messages, statusHistory, audit]
          properties:
            description:
              type: string
            replyContact:
              type: string
              nullable: true
            diagnostics:
              $ref: "#/components/schemas/SupportDiagnosticEnvelope"
              nullable: true
            messages:
              type: array
              items:
                $ref: "#/components/schemas/AdminSupportMessage"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/AdminSupportStatusHistory"
            audit:
              type: array
              items:
                $ref: "#/components/schemas/AdminSupportAudit"
            closedAt:
              type: string
              format: date-time
              nullable: true
...
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string